move OpenAICompletionProvider to providers location
