+++ crates/ai/src/providers/open_ai/mod.rs
@@ -1,2 +1,4 @@
+pub mod completion;
 pub mod model;
+pub use completion::*;
 pub use model::OpenAILanguageModel;
